# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Develop Experiments API
  version: 1.0.0
  description: |
    掲示板 API。スレッドとコメントの取得・投稿を提供します。

    一覧系はすべてキーセット (cursor) ページネーションです。
    OFFSET を使わないのは、深いページほど線形に遅くなることと、
    ページ送り中の挿入で行の重複・取りこぼしが起きるためです。

    カーソルは不透明トークンです。中身は実装詳細なので、
    クライアントは `nextCursor` をそのまま `cursor` に渡してください。
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
...
paths:
  /auth/google:
    get:
      operationId: startGoogleLogin
      tags: [auth]
      summary: Google へのログインを開始する
      description: |
        Google の認可エンドポイントへリダイレクトします。

        state / nonce / PKCE の code_verifier をサーバ側で生成し、
        短命な Cookie に保存します。CSRF とトークン差し替えを防ぐため、
        コールバックではこれらを必ず照合します。
      responses:
        '302':
          description: Google の認可エンドポイントへリダイレクト
          headers:
            Location:
              schema:
                type: string
              description: accounts.google.com の認可 URL
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/google/callback:
    get:
      operationId: googleLoginCallback
      tags: [auth]
      summary: Google からのコールバックを受ける
      description: |
        認可コードをトークンへ交換し、ID トークンを検証してから
        セッションを発行します。成功するとフロントエンドへリダイレクトし、
        セッション ID を HttpOnly Cookie で返します。

        退会済みの利用者が同じ Google アカウントで再ログインした場合は
        401 になります。行を復活させるかは未決です
        (docs/adr/0005-authentication.md)。

        同意画面で拒否された場合、Google は code を付けずに
        error だけを返します。この経路もフロントエンドへリダイレクトし、
        API のエラー JSON をブラウザに直接見せません。
      parameters:
        - name: state
          in: query
          required: true
          description: |
            開始時に発行した state。Cookie の値と照合します。
            拒否された場合も Google はこれを返すため、必須にしています
          schema:
            type: string
            maxLength: 512
        - name: code
          in: query
          required: false
          description: |
            Google が発行した認可コード。
            **同意画面で拒否されたときは付きません** (代わりに error が付きます)。
            必須にすると、拒否した利用者に検証エラーの JSON が直接見えてしまいます
          schema:
            type: string
            maxLength: 2048
        - name: error
          in: query
          required: false
          description: |
            拒否や設定不備のときに Google が返すエラー識別子
            (access_denied など)
          schema:
            type: string
            maxLength: 256
      responses:
        '302':
          description: フロントエンドへリダイレクト
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
              description: セッション ID (HttpOnly)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /threads:
    get:
      operationId: listThreads
      tags: [threads]
      summary: スレッド一覧の取得
      description: |
        スレッドを新しい順に、コメント数つきで取得します。

        コメント数は LEFT JOIN と COUNT(*) FILTER による単一クエリで集計しています。
        スレッドごとに COUNT を投げる実装 (N+1) を goroutine で並列化しても
        DB へのラウンドトリップ回数は減らないため、単一クエリのほうが有利です。

        `q` を渡すと、タイトルの中間一致で絞り込みます。
        **絞り込んでも並び順は新着順のまま**なので、`cursor` の使い方は変わりません
        (docs/adr/0012-search.md 決定 3)。

        `sort=popular` で閲覧数の多い順になります。
        **カーソルは並び順ごとに意味が変わります** ——
        `sort` を変えたら先頭から取り直してください。
        混ぜて渡した場合は 400 になります
        (docs/adr/0006-view-count-and-popularity.md)。
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createThread
      tags: [threads]
      summary: スレッドの作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThreadRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: |
            アイコンを指定したが未ログインです。
            **画像の投稿にはログインが要ります**
            (docs/adr/0007-image-storage.md の背景)。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: |
            指定されたアイコンが存在しないか、自分のものではありません。
            **403 ではなく 404** です —— 403 だと
            「その ID の画像が存在すること」自体が漏れます。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: |
            ストレージの設定が入っていないため、アイコンを指定できません。
            アイコンなしのスレッド作成は影響を受けません。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /threads/{threadId}:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
    get:
      operationId: getThread
      tags: [threads]
      summary: スレッドの取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteThread
      tags: [threads]
      summary: 自分のスレッドを削除する
      description: |
        **自分が立てたスレッドだけ**を削除できます
        (docs/adr/0005-authentication.md の権限モデル)。
        [ADR 0003](../docs/adr/0003-open-questions.md) の未決 #7
        (コメント削除 API を公開するか) は、認証が入ったこの時点で解けます。

        論理削除です。行は残り、コメントの参照整合性も保たれます。

        ### 匿名で立てたスレッドは削除できません

        `author_id` が NULL のスレッドは、**投稿者を特定する情報がない**ため
        本人であることを示せません。ログイン後に「さっき匿名で書いたやつ」を
        自分のものにする機能も作りません —— 許すと任意の匿名投稿を
        誰でも自分のものだと主張できます
        (docs/adr/0005-authentication.md 決定 2)。

        荒らしへの対処はモデレーターが行います
        (`POST /moderation/actions`。docs/adr/0011-moderation.md 決定 2)。

        ### コメントは消えません

        スレッドが論理削除されると一覧からも取得からも消えるため、
        コメントを個別に消して回る必要がありません
        (8 パーティションすべてに UPDATE を投げることにもなります)。

        ### 添付画像もストレージに残ります

        画像を消すのはモデレーションの操作です
        (docs/adr/0011-moderation.md 決定 5)。
        自分のスレッドを消しただけで実体まで消すと、
        誤操作の巻き戻しができなくなります。
      security:
        - sessionCookie: []
      responses:
        '204':
          description: 削除した。本文はありません
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            `code` は `PERMISSION_DENIED` です。原因は 3 つあります。

            - **他人のスレッド**
            - **匿名で立てられたスレッド** (`author_id` が NULL)
            - 状態変更メソッドの `Origin` / `Referer` が許可リストに無い
              (docs/adr/0013-http-defense.md 決定 1)

            **ここでは 404 に隠しません。** スレッドは誰でも
            `GET /threads/{threadId}` で読めるので、存在は公開情報です。
            隠す意味が無いうえ、404 にすると自分の投稿が消せないときに
            「消えたのか、権限が無いのか」を利用者が区別できません。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: スレッドが存在しないか、既に削除されています。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /threads/{threadId}/comments/{commentId}:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
      - $ref: '#/components/parameters/CommentId'
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: 自分のコメントを削除する
      description: |
        **自分が書いたコメントだけ**を削除できます。
        条件と理由はスレッドの削除と同じです。

        ### レス番号は空きます

        削除しても `seq` は再利用しません
        (docs/adr/0019-comment-concurrency.md 決定 5)。
        再利用すると、過去の `>>5` が別の投稿を指すようになります。

        ### スレッド ID がパスに要る理由

        `comments` は `thread_id` による HASH パーティションで、
        主キーが `(thread_id, id)` です。**コメント ID だけでは
        先頭列を絞れず、8 パーティションすべてを走査します。**
      security:
        - sessionCookie: []
      responses:
        '204':
          description: 削除した。本文はありません
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            `code` は `PERMISSION_DENIED` です。
            他人のコメント、匿名のコメント、または `Origin` の検証失敗です
            (スレッドの削除と同じ)。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: |
            コメントが存在しないか、既に削除されています。

            **指定したスレッドに属さないコメント ID も 404 です。**
            パーティションキーで絞るため、他スレッドの行には当たりません。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /threads/{threadId}/comments:
    parameters:
      - $ref: '#/components/parameters/ThreadId'
    get:
      operationId: listComments
      tags: [comments]
      summary: コメント一覧の取得
      description: |
        指定スレッドのコメントを新しい順に取得します。

        comments テーブルは thread_id による HASH パーティションなので、
        thread_id を等値指定するこのクエリでは partition pruning が効き、
        8 分割中 1 パーティションのみを走査します。
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createComment
      tags: [comments]
      summary: コメントの投稿
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: 投稿成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 親スレッドが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            同時更新が競合しました。再試行可能です。
            SERIALIZABLE 分離レベルでの直列化失敗 (SQLSTATE 40001) や
            デッドロック (40P01) がここに対応します。

            同じ `Idempotency-Key` の処理がまだ完了していない場合も
            ここに入ります (docs/adr/0015-idempotency.md)。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: |
            **同じ `Idempotency-Key` で、前回と違う内容が送られました。**

            前回の結果を黙って返すと、クライアントのバグが見えなくなります。
            キーを作り直して送り直してください
            (docs/adr/0015-idempotency.md)。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        ログイン時に発行される HttpOnly Cookie。

        **`security` を宣言したオペレーションだけが認証必須です。**
        匿名投稿を残す設計 (docs/adr/0005-authentication.md 決定 2) のため、
        スレッド作成やコメント投稿には宣言していません。
        OpenAPI の `security` は「必須」しか表現できず、
        「ログインしていれば投稿者を紐付ける」という任意認証は書けないためです。
        任意認証はハンドラ側で扱います。
  parameters:
    ThreadId:
      name: threadId
      in: path
      required: true
      description: スレッド ID
      schema:
        type: integer
        format: int64
        minimum: 1
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        次ページの取得位置を表す不透明トークン。
        直前のレスポンスの `nextCursor` をそのまま渡します。
        省略した場合と空文字を渡した場合は、どちらも先頭ページになります。

        中身はサーバ側の実装詳細なので、
        **クライアントは解釈も生成も改変もしないでください。**
        並び順を追加してもこの型は変わりません
        (docs/adr/0018-opaque-cursor.md)。
      schema:
        type: string
        # base64url (パディングなし) なので、この文字集合に収まる。
        # 空文字を許すのは、クライアントが ?cursor=${nextCursor ?? ''} と
        # 組み立てたときに初回ロードだけ 400 になるのを避けるため。
        # 長さの上限がないと、巨大な文字列を復号させられる。
        pattern: '^[A-Za-z0-9_-]*$'
        maxLength: 256
        example: eyJ2IjoxLCJpZCI6NDJ9
    Size:
      name: size
      in: query
      required: false
      description: |
        取得件数。省略時は 20。
        上限を設けないと、1 リクエストで全件走査させられてしまいます。
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: |
        並び順。

        - `new` (既定): 新着順 (id の降順)
        - `popular`: 閲覧数の多い順 (docs/adr/0006-view-count-and-popularity.md)

        **`q` との同時指定はできません** (400)。
        検索結果の並び順は新着順に固定してあり
        (docs/adr/0012-search.md 決定 3)、人気順を重ねると
        `title ILIKE` の GIN 索引と `(view_count, id)` の索引の
        どちらか一方しか使えません。どちらが有利かは検索語の珍しさで
        変わるため、**測ってから決めます** (Phase 4)。

        黙って新着順に落とさないのは、「人気順で並べたつもりの新着順」を
        返すと、利用者にもこちらにも間違いが見えないためです。
      schema:
        type: string
        enum: [new, popular]
        default: new
    SearchQuery:
      name: q
      in: query
      required: false
      description: |
        スレッドのタイトルに対する検索語 (中間一致)。

        - **対象はタイトルだけです。** コメント本文は検索できません
          (docs/adr/0012-search.md 決定 2)。
          コメントは 5 年で 1.8 億行規模になる見積もりで、
          そこに GIN 索引を張ると索引が本体に匹敵する大きさになります
        - **並び順は新着順のまま**で、関連度順ではありません (同 決定 3)。
          関連度順にするとカーソルがスコアになり、
          埋め込む値の意味が検索語ごとに変わります
        - **`%` と `_` はワイルドカードになりません。** サーバ側で
          エスケープするため、`100%` はその文字列として検索されます
        - **大文字と小文字は区別しません**
        - **前後の空白は取り除きます。空白だけの検索語は「指定なし」と同じ**
          扱いになり、絞り込みのない一覧を返します (400 にはしません)。
          検索欄を空のまま送信したフォームが 400 になるのを避けるためで、
          `cursor` に空文字を許しているのと同じ理由です

        検索語が 1〜2 文字だと索引の選択性が落ちます
        (`pg_trgm` は 3 文字単位のトライグラムに分解するため)。
        最低文字数の制限は設けていません —— 実測してから決めます
        (docs/adr/0012-search.md の罠)。
      schema:
        type: string
        # タイトルの上限と揃えています。これより長い検索語は
        # どのタイトルにも一致しないため、受け取っても無駄になります。
        # 空文字を許すのは、上の「指定なしと同じ」を成立させるためです。
        maxLength: 200
      example: PostgreSQL
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 対象が存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未ログイン、またはセッションが無効
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ServiceUnavailable:
      description: |
        認証プロバイダの設定が入っていないため、この経路は利用できません。
        設定が無くても API 自体は起動します
        (掲示板の閲覧と匿名投稿は認証に依存しないため)。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: サーバ内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Thread:
      type: object
      required: [id, title, commentCount, viewCount, author, icon, createdAt]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Go の並列処理を学ぶ部屋
        commentCount:
          type: integer
          format: int64
          description: 論理削除されていないコメントの件数
          example: 7
        viewCount:
          type: integer
          format: int64
          description: |
            閲覧数。

            **正確な値ではありません** (docs/adr/0006-view-count-and-popularity.md)。

            - **最新ではありません。** 計上はアプリのメモリ上で行い、
              一定間隔でまとめて反映するため、最大でその間隔ぶん遅れます
            - **欠落します。** プロセスが異常終了するとバッファ内の増分は消えます
            - **同一利用者の連打は一定時間まとめられます。** インスタンスを
              またいだ厳密性は保証しません

            表示用の指標であり、課金や順位の確定には使えません。
          example: 1284
        author:
          $ref: '#/components/schemas/Author'
        icon:
          allOf:
            - $ref: '#/components/schemas/Image'
          nullable: true
          description: |
            スレッドアイコン。**設定されていなければ `null`** です。
        createdAt:
          type: string
          format: date-time
          example: '2026-08-02T12:00:00Z'
...
    ThreadList:
      type: object
      required: [threads, nextCursor]
      properties:
        threads:
          type: array
          items:
            $ref: '#/components/schemas/Thread'
        nextCursor:
          $ref: '#/components/schemas/NextCursor'
...
    CreateThreadRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: DB 側の CHECK 制約 threads_title_length と同じ上限です
          example: 新しいスレッド
        iconImageId:
          type: string
          format: uuid
          description: |
            スレッドアイコンの画像 ID。`POST /images` で
            `kind=thread_icon` として上げたものを渡します。

            - **省略できます。** アイコンの無いスレッドが既定です
            - **自分がアップロードした画像だけ**を指定できます
              (他人の画像 ID は 404)
            - **未ログインでは指定できません** (画像の投稿にログインが要るため)
          example: 018f2c00-0000-7000-8000-000000000001
...